Fixed compiler warning on Windows.
git-svn-id: https://svn.opendtect.org/od/trunk@29844 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiTools/uidirectionalplot.cc
@@ -200,7 +200,7 @@ void uiDirectionalPlot::drawGrid()
 	const float ang = data_.angle( isect, 1 );
 	const float mathang = Angle::convert( data_.setup_.angletype_, ang,
 					      Angle::Rad );
-	uiLineItem* li = new uiLineItem( center_, mathang, radius_ );
+	uiLineItem* li = new uiLineItem( center_, mathang, (float) radius_ );
 	sectorlines_.add( li );
 	li->setPenStyle( setup_.sectorls_ );
     }
